configure recommended shared memory for selenium/standalone-chromium
- Fixes intermittent browser memory starvation during when running sytem tests in devcontainer. Error thrown during test is typically: Selenium::WebDriver::Error::InvalidSessionIdError: invalid session id

- From the docker-selenium readme:
"Why is --shm-size 2g necessary?
This is a known workaround to avoid the browser crashing inside a docker container, here are the documented issues for Chrome and Firefox. The shm size of 2gb is arbitrary but known to work well, your specific use case might need a different value, it is recommended to tune this value according to your needs."
https://github.com/SeleniumHQ/docker-selenium?tab=readme-ov-file#--shm-size2g